redo complete task query to only mark current assigned task as not assigned and if task is recurring, insert into recurring table and update due date in task table. If it isnt recurring, delete from task table and insert into completed task table.
Make sure description is disabled outside of edit mode

pass task instead of ID to Assign form to facilitate complete task

New Function CalculateNextDueDate for figuring out new due date for completed recurring tasks

Reset queries for active tasks when completing a task and refrest page

+++ components/CRUD_queries/CompleteTaskQuery.tsx
@@ -1,36 +1,70 @@
 import supabaseClient from "@/lib/supabaseClient";
 import { Task } from "@/types/Task";
+import { CalculateNextDueDate } from "../CalculateNextDueDate";
 
 const CompleteTaskQuery = async (taskData: Task, userId: string) => {
-  const { error: currentTaskError } = await supabaseClient
+  //Unassign task from user
+  const { error: OldActiveError } = await supabaseClient
     .from("user_current_task")
-    .delete()
-    .eq("user_id", userId);
+    .update({ is_assigned: false, is_current: false, action_at: new Date() })
+    .eq("user_id", userId)
+    .eq("is_current", true)
+    .eq("is_assigned", true);
 
-  if (currentTaskError) {
-    throw new Error(currentTaskError.message);
-  }
+  if (OldActiveError) throw new Error(OldActiveError.message);
 
-  const { error: completedTaskError } = await supabaseClient
-    .from("completed_tasks")
-    .insert({
-      task_id: taskData.id,
-      name: taskData.name,
-      description: taskData.description,
-      completed_by_user_id: userId,
-    });
+  //don't delete from main tasks table if it is a recurring task but reset the due date instead
+  if (taskData.is_recurring) {
+    const newDueDate = CalculateNextDueDate(taskData.recurring_type);
 
-  if (completedTaskError) {
-    throw new Error(completedTaskError.message);
-  }
+    console.log(JSON.stringify(newDueDate));
 
-  const { error: deleteTaskError } = await supabaseClient
-    .from("tasks")
-    .delete()
-    .eq("id", taskData.id);
+    const { error: recurringTaskError } = await supabaseClient
+      .from("tasks")
+      .update({ due_date: newDueDate, updated_at: new Date() })
+      .eq("created_by", userId)
+      .eq("id", taskData.id);
 
-  if (deleteTaskError) {
-    throw new Error(deleteTaskError.message);
+    if (recurringTaskError) throw new Error(recurringTaskError.message);
+
+    const { error: RecurringCompletedError } = await supabaseClient
+      .from("completed_recurring_tasks")
+      .insert({
+        task_id: taskData.id,
+        original_due_date: taskData.due_date,
+        completed_by: userId,
+        completed_at: new Date(),
+      });
+
+    if (RecurringCompletedError) {
+      throw new Error(RecurringCompletedError.message);
+    }
+  } else {
+    //Insert task data into completed table to keep data in logs
+    const { error: completedTaskError } = await supabaseClient
+      .from("completed_tasks")
+      .insert({
+        task_id: taskData.id,
+        name: taskData.name,
+        description: taskData.description,
+        priority: taskData.priority,
+        due_date: taskData.due_date,
+        completed_by: userId,
+      });
+
+    if (completedTaskError) {
+      throw new Error(completedTaskError.message);
+    }
+
+    const { error: deleteTaskError } = await supabaseClient
+      .from("tasks")
+      .delete()
+      .eq("id", taskData.id)
+      .eq("created_by", userId);
+
+    if (deleteTaskError) {
+      throw new Error(deleteTaskError.message);
+    }
   }
 };
